CORO/requirements.txt
fastapi
uvicorn[standard]
Pillow
requests
python-multipart
groq
python-dotenv
numpy
opencv-python-headless
